lambdas/satellite_palette/requirements.txt
Pillow==10.4.0
numpy==1.26.4
boto3==1.35.0
